# public/openapi.yaml
openapi: 3.0.3
info:
  title: Buggie API
  version: 1.0.0
  description: |
    Two HTTP APIs, described here together.

    - **The token API** (`/v1`) is for scripting against a workspace: listing, reading,
      filing and updating issues, and listing projects. It lives on the workspace's own
      subdomain and is authenticated by a personal access token made under
      Settings → Workspace → API tokens, with the abilities `read` and/or `write`.
      Tokens are for members of staff: a client's token is refused with 403. Requests
      are limited to 120 a minute per token.
    - **The reporter ingest API** (`/ingest`) is what the web widget and the native SDKs
      post bug reports to. It is unauthenticated and keyed by a project's public widget
      key. It lives on the central domain, `https://buggie.eu/api`, or a self-hosted
      install's own domain.

    Not covered: the inbound-mail webhook (`/mail/inbound`), which is Mailgun's callback
    rather than an API for Buggie's users; the webhooks Buggie sends to you; and the
    application's own screens.

    Read-then-write for the token API: issue and project reads include numeric ids
    (`project.id`, `status.id`, label `id`s). `GET /v1/projects/{key}` returns the
    project's statuses and the workspace's labels so a script can resolve names to
    ids before writing.

    Kept in step with the routes by `tests/Feature/OpenApiSpecTest.php`.
  license:
    name: AGPL-3.0
...
servers:
...
    description: >-
...
paths:
  /v1/issues:
    get:
      operationId: listIssues
      tags: [Issues]
      summary: List issues
      description: >-
        Issues the token's owner can see, filtered by `q` (the same query language as
        the filter bar, https://buggie.eu/docs/query-language). Ordered by priority,
        then most recently updated. Needs the `read` ability.
      security:
        - bearerAuth: []
      parameters:
        - name: q
          in: query
          required: false
          description: >-
            A query in Buggie's issue query language, e.g. `is:open project:web
            -label:wontfix`. The response echoes it back canonicalised under
            `meta.query`.
          schema:
            type: string
        - name: page
          in: query
          required: false
          description: The page to return, from 1.
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: per_page
          in: query
          required: false
          description: Issues per page. Values above 100 are capped at 100.
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
      responses:
        '200':
          description: A page of issues.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssuePage'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
    post:
      operationId: createIssue
      tags: [Issues]
      summary: File an issue
      description: Needs the `write` ability.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IssueCreate'
      responses:
        '201':
          description: The created issue.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssueResponse'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /v1/issues/{key}:
    parameters:
      - $ref: '#/components/parameters/IssueKey'
    get:
      operationId: getIssue
      tags: [Issues]
      summary: Read an issue
      description: Needs the `read` ability.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: The issue.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssueResponse'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
    patch:
      operationId: updateIssue
      tags: [Issues]
      summary: Update an issue
      description: Changes only the fields sent. Needs the `write` ability.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IssueUpdate'
      responses:
        '200':
          description: The updated issue.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssueResponse'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '429':
          $ref: '#/components/responses/TooManyRequests'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: >-
        A personal access token from Settings → Workspace → API tokens, for members of
        staff. It is shown once, works only in the workspace it was made in, and carries
        the abilities `read` (list and read issues and projects) and/or `write` (create
        and update issues).
  parameters:
    IssueKey:
      name: key
      in: path
      required: true
      description: The issue's key, e.g. `WEB-142`.
      schema:
        type: string
...
  responses:
    Unauthenticated:
      description: No token, or an expired or revoked one.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: >-
        Two different causes, told apart by `message`. Either the token lacks the
        ability this call needs, or its owner may not do this: mint a token with the
        right ability. Or its owner is not staff (`API tokens are for members of
        staff.`), which is final: a client's token, or one whose owner has since been
        moved to a client role, is refused on every call and no ability change helps.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: >-
        Not found, or not visible to the token's owner. A token presented on another
        workspace's subdomain also gets 404.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ValidationFailed:
      description: The request was understood and refused; `errors` says why, per field.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    TooManyRequests:
      description: Too many requests. Wait for `Retry-After` seconds where it is sent.
      headers:
        Retry-After:
          description: Seconds until another request is allowed.
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      required: [message]
      properties:
        message:
          type: string
        errors:
          type: object
          description: Field names, each with its validation messages.
          additionalProperties:
            type: array
            items:
              type: string
    IssueStatus:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
          example: In progress
        category:
          type: string
          enum: [backlog, unstarted, started, done, canceled]
        open:
          type: boolean
          description: Whether the status counts as open.
    IssueProject:
      type: object
      properties:
        id:
          type: integer
        key:
          type: string
          example: WEB
        slug:
          type: string
          example: web
        name:
          type: string
    PersonRef:
      type: object
      nullable: true
      properties:
        id:
          type: integer
        name:
          type: string
    LabelRef:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
    IssueSummary:
      type: object
      properties:
        key:
          type: string
          example: WEB-142
        title:
          type: string
        type:
          type: string
          enum: [bug, feature, task, question]
        priority:
          type: integer
          enum: [0, 1, 2, 3, 4]
          description: 0 none, 1 low, 2 medium, 3 high, 4 urgent.
        status:
          $ref: '#/components/schemas/IssueStatus'
        project:
          $ref: '#/components/schemas/IssueProject'
        assignee:
          $ref: '#/components/schemas/PersonRef'
        labels:
          type: array
          description: Labels on the issue, with ids for writes.
          items:
            $ref: '#/components/schemas/LabelRef'
        occurrences:
          type: integer
          description: How many times this bug has been reported, counting the first.
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Issue:
      allOf:
        - $ref: '#/components/schemas/IssueSummary'
        - type: object
          properties:
            description:
              type: object
              nullable: true
              description: >-
                A rich-text document in Tiptap's JSON form, e.g. `{"type": "doc",
                "content": [{"type": "paragraph", "content": [{"type": "text", "text":
                "Steps to reproduce"}]}]}`.
              additionalProperties: true
            reporter:
              $ref: '#/components/schemas/PersonRef'
            visibility:
              type: string
              enum: [internal, client]
            due_on:
              type: string
              format: date
              nullable: true
            closed_at:
              type: string
              format: date-time
              nullable: true
            estimate_minutes:
              type: integer
              nullable: true
            time_spent_minutes:
              type: integer
            custom_fields:
              description: >-
                Custom field values keyed by the field's key, e.g. `{"environment":
                "Production"}`. An issue with none serialises as an empty array, `[]`,
                rather than an empty object.
    IssuePage:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/IssueSummary'
        meta:
          type: object
          properties:
            page:
              type: integer
            per_page:
              type: integer
            total:
              type: integer
            last_page:
              type: integer
            query:
              type: string
              description: The `q` that was applied, canonicalised.
    IssueResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/Issue'
    IssueCreate:
      type: object
      required: [project_id, title]
      properties:
        project_id:
          type: integer
          description: >-
            The project's numeric id, for a project the token's owner can see. Read
            it from `GET /v1/projects` or `GET /v1/projects/{key}`.
        title:
          type: string
          maxLength: 255
        parent:
          type: string
          maxLength: 40
          nullable: true
          description: An issue key; the new issue is created as its subtask.
        description:
          type: object
          nullable: true
          description: A rich-text document in Tiptap's JSON form.
          additionalProperties: true
        type:
          type: string
          enum: [bug, feature, task, question]
          default: bug
        priority:
          type: integer
          enum: [0, 1, 2, 3, 4]
          default: 0
          description: 0 none, 1 low, 2 medium, 3 high, 4 urgent.
        status_id:
          type: integer
          nullable: true
          description: A status of the workspace. Defaults to the project's default status.
        assignee_id:
          type: integer
          nullable: true
          description: A member of staff. Clients cannot be assigned.
        visibility:
          type: string
          enum: [internal, client]
          default: internal
        labels:
          type: array
          description: Label ids.
          items:
            type: integer
        custom_fields:
          type: object
          description: Values keyed by the field's key, checked against the project's fields.
          additionalProperties: true
    IssueUpdate:
      type: object
      properties:
        title:
          type: string
          maxLength: 255
        description:
          type: object
          nullable: true
          description: A rich-text document in Tiptap's JSON form.
          additionalProperties: true
        type:
          type: string
          enum: [bug, feature, task, question]
        priority:
          type: integer
          enum: [0, 1, 2, 3, 4]
          description: 0 none, 1 low, 2 medium, 3 high, 4 urgent.
        visibility:
          type: string
          enum: [internal, client]
        client_audience:
          type: string
          enum: [default, project, specific]
          description: Which clients see a client-visible issue.
        board_pinned:
          type: boolean
          description: A pinned card stays on the board whatever the filter says.
        client_share_ids:
          type: array
          maxItems: 100
          description: >-
            The clients an issue is shared with, by user id, when `client_audience` is
            `specific`.
          items:
            type: integer
        start_on:
          type: string
          format: date
          nullable: true
        due_on:
          type: string
          format: date
          nullable: true
        version_id:
          type: integer
          nullable: true
          description: A release of the issue's project.
        phase_id:
          type: integer
          nullable: true
          description: A phase of the issue's project.
        status_id:
          type: integer
          description: A status of the workspace.
        assignee_id:
          type: integer
          nullable: true
          description: A member of staff. Clients cannot be assigned.
        labels:
          type: array
          description: Label ids. Replaces the issue's labels.
          items:
            type: integer
        custom_fields:
          type: object
          description: Values keyed by the field's key, checked against the project's fields.
          additionalProperties: true
...